LineBot/requirements.txt
line-bot-sdk
flask
configparser
openai
azure-iot-hub